Update quest api, and fixed parsing crash for mission's reward
BadgeData model class structure is shared among reward type of response back from json response. Thus marked it (for now) as such [SharedModel] in the comment.
Added FilteredParam, and Url for model classes.

Fixed parsing of mission's reward by skipping value if we need to too.

Note the Condition class which is base and generic for other condition that can be set in Playbasis dashboard. Thus note the serialized names there that are different from member variables.